backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
faiss-cpu>=1.7.0
joblib>=1.3.0
python-multipart>=0.0.6
pytest>=8.0.0
